@@ -1,7 +1,7 @@
 package com.company.swiftchatsupportlibrary
 
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.widget.Button
 import com.swift.chat.library.SwiftChatActivity
 
